requirements.txt
celery==4.4.7
Django==3.1.7
django-debug-toolbar==3.2
django-environ==0.4.5
djangorestframework==3.12.2
django-rest-knox==4.1.0
redis==3.5.3
drf-yasg==1.20.0
psycopg2==2.9.6
flower==0.9.7
gunicorn==20.0.4
django-redis==4.12.1
django-parler==2.2
sorl-thumbnail==12.7.0
django-filter==2.4.0
django-cachalot==2.4.1
django-mptt==0.12.0
django-fieldsignals==0.6.0
django-solo==1.1.5
django-celery-email==3.0.0
django-searchable-select==1.5.0
django-parler-rest==2.1
six==1.16.0
Pillow==9.5.0
django-ckeditor==6.2.0
dadata==20.7.0
django-colorfield==0.4.2
coreapi>=2.3.3
pytz>=2021.1
requests>=2.18.4
pyparsing>=2.4.7
django-autoslug>=1.9.8
Unidecode>=1.3.2
django-select2==7.9.0
django-clearcache==1.2.1
social-auth-app-django==5.0.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.35
django-cors-headers==3.10.0
